+++ src/Blockchain/exports/Get-AzBlockchainMember.ps1
@@ -1,0 +1,195 @@
+
+# ----------------------------------------------------------------------------------
+#
+# Copyright Microsoft Corporation
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+# http://www.apache.org/licenses/LICENSE-2.0
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ----------------------------------------------------------------------------------
+
+<#
+.Synopsis
+Get details about a blockchain member.
+.Description
+Get details about a blockchain member.
+.Example
+PS C:\> Get-AzBlockchainMember 
+
+Location Name               Type
+-------- ----               ----
+eastus   blockchainmember01 Microsoft.Blockchain/blockchainMembers
+eastus   myblockchain       Microsoft.Blockchain/blockchainMembers
+eastus   myblockchaine0f3ol Microsoft.Blockchain/blockchainMembers
+eastus   myblockchainuvbqdl Microsoft.Blockchain/blockchainMembers
+.Example
+PS C:\> Get-AzBlockchainMember -ResourceGroupName testgroup
+
+Location Name       Type
+-------- ----       ----
+eastus   dolauli001 Microsoft.Blockchain/blockchainMembers
+eastus   dolauli002 Microsoft.Blockchain/blockchainMembers
+.Example
+PS C:\> Get-AzBlockchainMember -Name dolauli001 -ResourceGroupName testgroup
+
+Location Name       Type
+-------- ----       ----
+eastus   dolauli001 Microsoft.Blockchain/blockchainMembers
+.Example
+PS C:\> $member = Get-AzBlockchainMember -ResourceGroupName blockchain-rg -Name myblockchaine0f3ol
+PS C:\> Get-AzBlockchainMember -InputObject $membe 
+
+Location Name               Type
+-------- ----               ----
+eastus   myblockchaine0f3ol Microsoft.Blockchain/blockchainMembers
+
+.Inputs
+Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Models.IBlockchainIdentity
+.Outputs
+Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Models.Api20180601Preview.IBlockchainMember
+.Notes
+COMPLEX PARAMETER PROPERTIES
+
+To create the parameters described below, construct a hash table containing the appropriate properties. For information on hash tables, run Get-Help about_Hash_Tables.
+
+INPUTOBJECT <IBlockchainIdentity>: Identity Parameter
+  [BlockchainMemberName <String>]: Blockchain member name.
+  [Id <String>]: Resource identity path
+  [Location <String>]: Location Name.
+  [OperationId <String>]: Operation Id.
+  [ResourceGroupName <String>]: The name of the resource group that contains the resource. You can obtain this value from the Azure Resource Manager API or the portal.
+  [SubscriptionId <String>]: Gets the subscription Id which uniquely identifies the Microsoft Azure subscription. The subscription ID is part of the URI for every service call.
+  [TransactionNodeName <String>]: Transaction node name.
+.Link
+https://docs.microsoft.com/en-us/powershell/module/az.blockchain/get-azblockchainmember
+#>
+function Get-AzBlockchainMember {
+[OutputType([Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Models.Api20180601Preview.IBlockchainMember])]
+[CmdletBinding(DefaultParameterSetName='List1', PositionalBinding=$false)]
+param(
+    [Parameter(ParameterSetName='Get', Mandatory)]
+    [Alias('BlockchainMemberName')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Path')]
+    [System.String]
+    # Blockchain member name.
+    ${Name},
+
+    [Parameter(ParameterSetName='Get', Mandatory)]
+    [Parameter(ParameterSetName='List', Mandatory)]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Path')]
+    [System.String]
+    # The name of the resource group that contains the resource.
+    # You can obtain this value from the Azure Resource Manager API or the portal.
+    ${ResourceGroupName},
+
+    [Parameter(ParameterSetName='Get')]
+    [Parameter(ParameterSetName='List')]
+    [Parameter(ParameterSetName='List1')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Path')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Runtime.DefaultInfo(Script='(Get-AzContext).Subscription.Id')]
+    [System.String[]]
+    # Gets the subscription Id which uniquely identifies the Microsoft Azure subscription.
+    # The subscription ID is part of the URI for every service call.
+    ${SubscriptionId},
+
+    [Parameter(ParameterSetName='GetViaIdentity', Mandatory, ValueFromPipeline)]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Path')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Models.IBlockchainIdentity]
+    # Identity Parameter
+    # To construct, see NOTES section for INPUTOBJECT properties and create a hash table.
+    ${InputObject},
+
+    [Parameter()]
+    [Alias('AzureRMContext', 'AzureCredential')]
+    [ValidateNotNull()]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Azure')]
+    [System.Management.Automation.PSObject]
+    # The credentials, account, tenant, and subscription used for communication with Azure.
+    ${DefaultProfile},
+
+    [Parameter(DontShow)]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Runtime')]
+    [System.Management.Automation.SwitchParameter]
+    # Wait for .NET debugger to attach
+    ${Break},
+
+    [Parameter(DontShow)]
+    [ValidateNotNull()]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Runtime')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Runtime.SendAsyncStep[]]
+    # SendAsync Pipeline Steps to be appended to the front of the pipeline
+    ${HttpPipelineAppend},
+
+    [Parameter(DontShow)]
+    [ValidateNotNull()]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Runtime')]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Runtime.SendAsyncStep[]]
+    # SendAsync Pipeline Steps to be prepended to the front of the pipeline
+    ${HttpPipelinePrepend},
+
+    [Parameter(DontShow)]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Runtime')]
+    [System.Uri]
+    # The URI for the proxy server to use
+    ${Proxy},
+
+    [Parameter(DontShow)]
+    [ValidateNotNull()]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Runtime')]
+    [System.Management.Automation.PSCredential]
+    # Credentials for a proxy server to use for the remote call
+    ${ProxyCredential},
+
+    [Parameter(DontShow)]
+    [Microsoft.Azure.PowerShell.Cmdlets.Blockchain.Category('Runtime')]
+    [System.Management.Automation.SwitchParameter]
+    # Use the default credentials for the proxy
+    ${ProxyUseDefaultCredentials}
+)
+
+begin {
+    try {
+        $outBuffer = $null
+        if ($PSBoundParameters.TryGetValue('OutBuffer', [ref]$outBuffer)) {
+            $PSBoundParameters['OutBuffer'] = 1
+        }
+        $parameterSet = $PSCmdlet.ParameterSetName
+        $mapping = @{
+            Get = 'Az.Blockchain.private\Get-AzBlockchainMember_Get';
+            GetViaIdentity = 'Az.Blockchain.private\Get-AzBlockchainMember_GetViaIdentity';
+            List = 'Az.Blockchain.private\Get-AzBlockchainMember_List';
+            List1 = 'Az.Blockchain.private\Get-AzBlockchainMember_List1';
+        }
+        if (('Get', 'List', 'List1') -contains $parameterSet -and -not $PSBoundParameters.ContainsKey('SubscriptionId')) {
+            $PSBoundParameters['SubscriptionId'] = (Get-AzContext).Subscription.Id
+        }
+        $wrappedCmd = $ExecutionContext.InvokeCommand.GetCommand(($mapping[$parameterSet]), [System.Management.Automation.CommandTypes]::Cmdlet)
+        $scriptCmd = {& $wrappedCmd @PSBoundParameters}
+        $steppablePipeline = $scriptCmd.GetSteppablePipeline($MyInvocation.CommandOrigin)
+        $steppablePipeline.Begin($PSCmdlet)
+    } catch {
+        throw
+    }
+}
+
+process {
+    try {
+        $steppablePipeline.Process($_)
+    } catch {
+        throw
+    }
+}
+
+end {
+    try {
+        $steppablePipeline.End()
+    } catch {
+        throw
+    }
+}
+}