+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 1.8.1
+* Updated assembly version of data factory V2 cmdlets
+
 ## Version 1.8.0
 * Supported CRUD of data flow runtime properties in Managed IR.