+++ src/OperationalInsights/OperationalInsights/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 ## Upcoming Release
 
+## Version 2.1.0
+* Upgraded SDK to 0.21.0
+* Added optional parameters to 
+    - `New-AzOperationalInsightsSavedSearch`
+    - `Set-AzOperationalInsightsSavedSearch`
+
 ## Version 2.0.0
 * Updated legacy code to apply new generated SDK
 * Deleted cmdlets due to deprecated APIs